ml1-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
requests
pillow
